Move integration tests from test/functional/ to test/integration/ (#10800)
