+++ extensions/amp-viewer-integration/0.1/test/integration/test-amp-viewer-integration.js
@@ -14,19 +14,19 @@
  * limitations under the License.
  */
 
-import {AmpViewerIntegration} from '../amp-viewer-integration';
+import {AmpViewerIntegration} from '../../amp-viewer-integration';
 import {
   Messaging,
   WindowPortEmulator,
   parseMessage,
-} from '../messaging/messaging';
-import {ViewerForTesting} from './viewer-for-testing';
-import {getSourceUrl} from '../../../../src/url';
+} from '../../messaging/messaging';
+import {ViewerForTesting} from '../viewer-for-testing';
+import {getSourceUrl} from '../../../../../src/url';
 
 
 describes.sandboxed('AmpViewerIntegration', {}, () => {
   const ampDocSrc = '/test/fixtures/served/ampdoc-with-messaging.html';
-  describe('Handshake', function() {
+  describe.configure().skipSauceLabs().run('Handshake', function() {
     let viewerEl;
     let viewer;
     let ampDocUrl;
@@ -139,7 +139,8 @@ describes.sandboxed('AmpViewerIntegration', {}, () => {
     });
   });
 
-  describe('Unit Tests for messaging.js', () => {
+  describe.configure().skipSauceLabs().run(`Unit Tests for
+      messaging.js`, () => {
     const viewerOrigin = 'http://localhost:9876';
     const requestProcessor = function() {
       return Promise.resolve({});
