+++ extensions/amp-viewer-integration/0.1/test/integration/test-amp-webview-viewer-integration.js
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-import {AmpViewerIntegration} from '../amp-viewer-integration';
-import {WebviewViewerForTesting} from './webview-viewer-for-testing.js';
+import {AmpViewerIntegration} from '../../amp-viewer-integration';
+import {WebviewViewerForTesting} from '../webview-viewer-for-testing.js';
 
 
 describes.sandboxed('AmpWebviewViewerIntegration', {}, () => {
   const ampDocSrc = '/test/fixtures/served/ampdoc-with-messaging.html';
-  describe('Handshake', function() {
+  describe.configure().skipSauceLabs().run('Handshake', function() {
     let viewerEl;
     let viewer;
