+++ extensions/amp-viewer-integration/0.1/test/integration/test-viewer-initiated-handshake.js
@@ -16,12 +16,12 @@
 
 import {
     WebviewViewerForTesting,
-} from './viewer-initiated-handshake-viewer-for-testing';
+} from '../viewer-initiated-handshake-viewer-for-testing';
 
 
 describes.sandboxed('AmpWebviewViewerIntegration', {}, () => {
   const ampDocSrc = '/test/fixtures/served/ampdoc-with-messaging.html';
-  describe('Handshake', function() {
+  describe.configure().skipSauceLabs().run('Handshake', function() {
     let viewerEl;
     let viewer;
